Introduce XRHandSphereModel
